+++ www/Site/Library/content_widgets/Page/Page.php
@@ -25,7 +25,7 @@ class Page extends AutoWidgetList2
             $cond['where'] = array('all', array(
                 array('attr', 'is_hidden', '=', $this->_input['REQUEST']['object']->attr('is_hidden')),
                 array('attr', 'is_draft', '=', 0),
-                array('attr', 'is_mandatory', '=', 1),
+                array('attr', 'is_property', '=', 1),
                 array('attr', 'diff', '!=', Entity::DIFF_ADD)
             ));
         }