+++ www/Site/Library/menus/NextPrevNavigation/NextPrevNavigation.php
@@ -8,6 +8,7 @@
 
 namespace Library\menus\NextPrevNavigation;
 
+use Boolive\data\Entity;
 use Library\views\Widget\Widget,
     Boolive\values\Rule,
     Boolive\values\Check,
@@ -58,6 +59,10 @@ class NextPrevNavigation extends Widget
         $next = $object->parent()->find(array(
                 'where' => array(
                     array('attr', 'order', '>', $object->order()),
+                    array('attr', 'is_hidden', '=', 0),
+                    array('attr', 'is_draft', '=', 0),
+                    array('attr', 'is_property', '=', 0),
+                    array('attr', 'diff', '!=', Entity::DIFF_ADD),
                     array('is', $object_types)
                 ),
                 'order' => array(
@@ -70,6 +75,10 @@ class NextPrevNavigation extends Widget
         $prev = $object->parent()->find(array(
             'where' => array(
                     array('attr', 'order', '<', $object->order()),
+                    array('attr', 'is_hidden', '=', 0),
+                    array('attr', 'is_draft', '=', 0),
+                    array('attr', 'is_property', '=', 0),
+                    array('attr', 'diff', '!=', Entity::DIFF_ADD),
                     array('is', $object_types)
                 ),
                 'order' => array(